refactor: Encapsulate timetable inside TimetableSnapshot.
This call to the updates listener exposed the internal map of timetable. The callee did only use the map to fetch data, so we could easily change the callback from a map to a function.

NOTE! There is a functional change in this commit! The `TimetableSnapshot#clearTimetables(...)` uses the TripPattern.id.feedScope to filter, not the TripPattern.route.id.feedScope. These should have tha same value, but it is possible to mess up here.

```Java
private boolean clearTimetables(String feedId) {
  return timetables.keySet().removeIf(id -> feedId.equals(id.getFeedId()));
}
```