refactor: Remove building arrival-listener from McStopArrivals
The McStopArrivals is a critical class, building the listeners is just noice and can be done in a mapper and injected.

This also remove StopArrivalParetoSet - witch only existed to take part of the listeners construction.

A tiny business rule is no enforced by the the mapper, not ideal, but overall much better than it was. The mapper is responsible for the order the listeners are added.