fix: build and deploy debug client also when graphql schema changes
The debug client now also depends on the graphql schema to build certain ui elements dynamically.
Parts of this happens runtime but it also has build-time dependencies. We therefore need to
rebuild and release a new version of the debug client whenever the graphql schema changes.

This commit updates the path filters in the github workflow trigger to accomplish this.